modularizing code to implement database connection later on and adding queries in comments.
