+++ server/models/author.js
@@ -1,0 +1,7 @@
+const authors = [
+    { name: 'Patrick Rothfuss', age: 44, id: '1' },
+    { name: 'Brandon Sanderson', age: 42, id: '2' },
+    { name: 'Terry Pratchett', age: 66, id: '3' }
+];
+module.exports = authors
+
